+++ include/unity/scopes/internal/RuntimeImpl.h
@@ -72,6 +72,7 @@ private:
     bool confined() const;
     std::string confinement_type() const;
     std::string find_cache_dir() const;
+    std::string find_app_dir() const;
     std::string find_tmp_dir() const;
 
     bool destroyed_;
@@ -85,7 +86,8 @@ private:
     mutable std::string ss_registry_identity_;
     int reap_expiry_;
     int reap_interval_;
-    std::string data_dir_;
+    std::string cache_dir_;
+    std::string app_dir_;
     std::string config_dir_;
     std::string tmp_dir_;
     mutable Reaper::SPtr reply_reaper_;